Deprecate yumpkg and replace it with yumpkg5 (renamed)
